Initial commit, first couple hours.
Config, database access, twitter oauth, retrieving and inserting / updating
tweets done. No query awareness yet.

Note that I'm not putting the twitter credentials into the repo. These
will be provided over email; it's not exactly a big deal for a throwaway
project like this, but I'd rather not lose access in the middle of a demo
or something.